Move tests and test models out of the package
