--- phonenumber_field/models.py
@@ -1,1 +1,0 @@
-# -*- coding: utf-8 -*-
