+++ setup.cfg
@@ -29,8 +29,6 @@ python_files = tests.py test_*.py *_tests.py
 [coverage:run]
 source = phonenumber_field
 branch = True
-omit =
-    phonenumber_field/tests.py
 
 [coverage:report]
 exclude_lines =
